dashboard/backend/requirements.txt
asgiref==3.8.1
Django==4.2.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
et-xmlfile==1.1.0
numpy==2.0.1
openpyxl==3.1.5
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
setuptools==72.2.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
wheel==0.44.0
